BC-12022 - Split interfaces importing `FileRecord` in separate folder `domain/contract`
