+++ src/modules/files-storage/domain/contract/file-record-with-status.interface.ts
@@ -1,0 +1,7 @@
+import { type FileRecord } from '../file-record.do';
+import { type FileRecordStatus } from '../interface';
+
+export interface FileRecordWithStatus {
+	fileRecord: FileRecord;
+	status: FileRecordStatus;
+}
