+++ src/modules/files-storage/domain/contract/file-record.repo.interface.ts
@@ -1,9 +1,9 @@
 import { type FindOptions } from '@shared/domain/interface';
 import { type Counted, type EntityId } from '@shared/domain/types';
 import { type FileRecord } from '../file-record.do';
+import { type StorageLocation } from '../interface';
 import { type StorageType } from '../storage-paths.const';
 import { type ParentStatistic } from '../vo';
-import { type StorageLocation } from './storage-location.enum';
 
 export interface FileRecordRepo {
 	findOneById(id: EntityId): Promise<FileRecord>;
