+++ src/modules/files-storage/domain/contract/index.ts
@@ -1,0 +1,3 @@
+export { FileRecordWithStatus } from './file-record-with-status.interface';
+export { FILE_RECORD_REPO, FileRecordRepo } from './file-record.repo.interface';
+export { PreviewFileParams } from './preview-file-params.interface';
