+++ src/modules/files-storage/domain/contract/preview-file-params.interface.ts
@@ -1,0 +1,10 @@
+import { type FileRecord } from '../file-record.do';
+import { type PreviewInfo } from '../interface';
+
+export interface PreviewFileParams {
+	fileRecord: FileRecord;
+	previewParams: PreviewInfo;
+	hash: string;
+	format: string;
+	bytesRange?: string;
+}
