+++ src/modules/files-storage/domain/interface/collabora-editability-status.interface.ts
@@ -1,0 +1,4 @@
+export interface CollaboraEditabilityStatus {
+	isCollaboraEditable: boolean;
+	exceedsCollaboraEditableFileSize: boolean;
+}
