+++ src/modules/files-storage/domain/interface/copy-file-result.interface.ts
@@ -1,0 +1,5 @@
+export interface CopyFileResult {
+	id?: string;
+	sourceId: string;
+	name: string;
+}
