+++ src/modules/files-storage/domain/interface/file-record-parent-type.enum.ts
