+++ src/modules/files-storage/domain/interface/file-record-status.interface.ts
@@ -1,19 +1,9 @@
-import { type FileRecord, type PreviewStatus } from '../file-record.do';
 import { type ScanStatus } from '../vo';
+import { type PreviewStatus } from './preview-status.enum';
 
 export interface FileRecordStatus {
 	scanStatus: ScanStatus;
 	previewStatus: PreviewStatus;
-	isCollaboraEditable: boolean;
-	exceedsCollaboraEditableFileSize: boolean;
-}
-
-export interface FileRecordWithStatus {
-	fileRecord: FileRecord;
-	status: FileRecordStatus;
-}
-
-export interface CollaboraEditabilityStatus {
 	isCollaboraEditable: boolean;
 	exceedsCollaboraEditableFileSize: boolean;
 }
