+++ src/modules/files-storage/domain/interface/get-file-response.interface.ts
@@ -1,0 +1,10 @@
+import type { Readable } from 'node:stream';
+
+export interface GetFileResponse {
+	data: Readable;
+	etag?: string;
+	contentType?: string;
+	contentLength?: number;
+	contentRange?: string;
+	name: string;
+}
