+++ src/modules/files-storage/domain/interface/index.ts
@@ -1,8 +1,17 @@
-export * from './document-type.enum';
-export * from './file-record-status.interface';
-export * from './file-record.repo.interface';
-export * from './file-storage-parent-type.enum';
-export * from './interfaces';
-export * from './parent-info.interface';
-export * from './preview-width.enum';
-export * from './storage-location.enum';
+export { CollaboraEditabilityStatus } from './collabora-editability-status.interface';
+export { CollaboraMimeTypes } from './collabora-mime-types.enum';
+export { CopyFileResult } from './copy-file-result.interface';
+export { DocumentType } from './document-type.enum';
+export { FileRecordParentType } from './file-record-parent-type.enum';
+export { FileRecordProps } from './file-record-props.interface';
+export { FileRecordStatus } from './file-record-status.interface';
+export { GetFileResponse } from './get-file-response.interface';
+export { ParentInfo } from './parent-info.interface';
+export { ParentReference } from './parent-reference.interface';
+export { PreviewInfo } from './preview-info.interface';
+export { PreviewOutputMimeTypes } from './preview-output-mime-types.enum';
+export { PreviewStatus } from './preview-status.enum';
+export { PreviewWidth } from './preview-width.enum';
+export { StorageLocationParams } from './storage-location-params.interface';
+export { StorageLocation } from './storage-location.enum';
+export { StorageReference } from './storage-reference.interface';
