+++ src/modules/files-storage/domain/interface/parent-reference.interface.ts
@@ -1,0 +1,7 @@
+import { type EntityId } from '@shared/domain/types';
+import { type FileRecordParentType } from './file-record-parent-type.enum';
+
+export interface ParentReference {
+	parentId: EntityId;
+	parentType: FileRecordParentType;
+}
