+++ src/modules/files-storage/domain/interface/preview-info.interface.ts
@@ -1,0 +1,8 @@
+import { type PreviewOutputMimeTypes } from './preview-output-mime-types.enum';
+import { type PreviewWidth } from './preview-width.enum';
+
+export interface PreviewInfo {
+	outputFormat?: PreviewOutputMimeTypes;
+	width?: PreviewWidth;
+	forceUpdate?: boolean;
+}
