+++ src/modules/files-storage/domain/interface/preview-output-mime-types.enum.ts
@@ -1,0 +1,3 @@
+export enum PreviewOutputMimeTypes {
+	IMAGE_WEBP = 'image/webp',
+}
