+++ src/modules/files-storage/domain/service/files-storage-copy.service.spec.ts
@@ -7,14 +7,15 @@ import { ObjectId } from '@mikro-orm/mongodb';
 import { ForbiddenException } from '@nestjs/common';
 import { Test, type TestingModule } from '@nestjs/testing';
 import {
-	FILE_STORAGE_CONFIG_TOKEN,
-	FILES_STORAGE_S3_CONNECTION,
-	type FileStorageConfig,
+    FILE_STORAGE_CONFIG_TOKEN,
+    FILES_STORAGE_S3_CONNECTION,
+    type FileStorageConfig,
 } from '../../files-storage.config';
 import { fileRecordTestFactory, ParentInfoTestFactory } from '../../testing';
+import { FILE_RECORD_REPO, type FileRecordRepo } from '../contract';
 import { ErrorType } from '../error';
 import { FilePathFactory, FileRecordFactory } from '../factory';
-import { type CopyFileResult, FILE_RECORD_REPO, type FileRecordRepo } from '../interface';
+import { type CopyFileResult } from '../interface';
 import { ScanStatus } from '../vo';
 import { FilesStorageService } from './files-storage.service';
