+++ src/modules/files-storage/domain/service/files-storage-remove-creator.service.spec.ts
@@ -10,8 +10,8 @@ import {
 	type FileStorageConfig,
 } from '../../files-storage.config';
 import { fileRecordTestFactory } from '../../testing';
+import { FILE_RECORD_REPO, type FileRecordRepo } from '../contract';
 import { FileRecord } from '../file-record.do';
-import { FILE_RECORD_REPO, type FileRecordRepo } from '../interface';
 import { FilesStorageService } from './files-storage.service';
 
 describe('FilesStorageService delete methods', () => {
