+++ src/modules/files-storage/testing/file-record-with-status.test.factory.ts
@@ -1,4 +1,4 @@
-import { type FileRecordWithStatus } from '../domain/interface/file-record-status.interface';
+import { type FileRecordWithStatus } from '../domain/contract/file-record-with-status.interface';
 import { fileRecordStatusTestFactory } from './file-record-status.test.factory';
 import { fileRecordTestFactory } from './file-record.test.factory';
